+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### AppCenter
 
+* **[Feature]** Allow users to set userId for crashes and error logs.
 * **[Fix]** Do not delete old logs when trying to add a log larger than the maximum storage capacity.
 * **[Fix]** Fix error detection of `setMaxStorageSize` API if database uses custom page size.
 * **[Fix]** Fix minimum storage size verification to match minimum possible value.