+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * **[Fix]** Do not delete old logs when trying to add a log larger than the maximum storage capacity.
 * **[Fix]** Fix error detection of `setMaxStorageSize` API if database uses custom page size.
 * **[Fix]** Fix minimum storage size verification to match minimum possible value.
+* **[Fix]** Fix logs duplication on unstable network.
 
 ### AppCenterCrashes